Test migrated to TableTest. Correct Javadoc and code style.
